+++ @navikt/core/tokens/darkside/tokens/colors/semantic-roles/success.tokens.ts
@@ -95,12 +95,12 @@ export const successSemanticTokenConfig = {
       comment:
         "Standard farge for tekst og ikoner med rollen success. Godkjent på alle bakgrunner unntatt strong.",
     },
-    "success-icon": {
+    "success-decoration": {
       value: "{ax.success.600.value}",
       type: "color",
       group: "text.success",
       comment:
-        "En farge som kun brukes på ikke-tekstlig innhold (ikoner og andre grafiske elementer). Ikke godkjent på tekst altså.",
+        "En farge som kun brukes på ikke-tekstlig innhold (ikoner og andre grafiske elementer). **Ikke godkjent på tekst altså**.",
     },
     "success-contrast": {
       value: "{ax.neutral.000.value}",