+++ aksel.nav.no/website/app/_sanity/queries.ts
@@ -1,4 +1,4 @@
-import { defineQuery, groq } from "next-sanity";
+import { defineQuery } from "next-sanity";
 import { contributorsAll, destructureBlocks } from "@/sanity/queries";
 
 const DESIGNSYSTEM_TYPES = `"komponent_artikkel", "ds_artikkel", "templates_artikkel"`;
@@ -51,19 +51,6 @@ const DESIGNSYSTEM_TEMPLATES_LANDINGPAGE_QUERY = defineQuery(
   `*[_type == "templates_landingsside"][0]`,
 );
 
-const searchContent = groq`{
-  heading,
-  "slug": slug.current,
-  "tema": undertema[]->tema->title,
-  ingress,
-  status,
-  _type,
-  "intro": pt::text(intro.body),
-  content,
-  publishedAt,
-  seo
-}`;
-
 const GLOBAL_SEARCH_QUERY_ALL = defineQuery(
   `*[_type in ["komponent_artikkel",
   "ds_artikkel",
@@ -71,7 +58,18 @@ const GLOBAL_SEARCH_QUERY_ALL = defineQuery(
   "aksel_blogg",
   "aksel_prinsipp",
   "aksel_standalone",
-  "templates_artikkel"]]${searchContent}`,
+  "templates_artikkel"]]{
+    heading,
+    "slug": slug.current,
+    "tema": undertema[]->tema->title,
+    ingress,
+    status,
+    _type,
+    "intro": pt::text(intro.body),
+    content,
+    publishedAt,
+    seo
+}`,
 );
 
 const KOMPONENT_BY_SLUG_QUERY =
@@ -157,6 +155,7 @@ const SLUG_BY_TYPE_QUERY = defineQuery(`
 const GOD_PRAKSIS_ALL_TEMA_QUERY =
   defineQuery(`*[_type == "gp.tema"] | order(lower(title)){
   title,
+  _updatedAt,
   description,
   pictogram,
   "slug": slug.current,
@@ -262,30 +261,48 @@ const DOCUMENT_BY_ID_FOR_SLACK_QUERY = defineQuery(`*[_id == $id][0]{
       "contacts": undertema[]->tema->contacts[]->email
     }`);
 
+/* --------------------------------- Sitemap -------------------------------- */
+const SITEMAP_LANDINGPAGES_QUERY = defineQuery(`
+{
+      "frontpage": *[_type == "aksel_forside"][0]._updatedAt,
+      "godpraksis": *[_type == "godpraksis_landingsside"][0]._updatedAt,
+      "blogg": *[_type == "blogg_landingsside"][0]._updatedAt,
+}
+  `);
+
+const SITEMAP_ARTICLES_BY_TYPE_QUERY = defineQuery(`
+  *[_type in $doctypes]{
+    "slug": slug.current,
+    _updatedAt
+  }
+  `);
+
 /* --------------------------------- Exports -------------------------------- */
 export {
-  DESIGNSYSTEM_SIDEBAR_QUERY,
-  DESIGNSYSTEM_OVERVIEW_PAGES_QUERY,
-  DESIGNSYSTEM_KOMPONENTER_LANDINGPAGE_QUERY,
-  DESIGNSYSTEM_GRUNNLEGGENDE_LANDINGPAGE_QUERY,
-  DESIGNSYSTEM_TEMPLATES_LANDINGPAGE_QUERY,
-  GLOBAL_SEARCH_QUERY_ALL,
-  KOMPONENT_BY_SLUG_QUERY,
-  DESIGNSYSTEM_OVERVIEW_BY_CATEGORY_QUERY,
-  TOC_BY_SLUG_QUERY,
-  SLUG_BY_TYPE_QUERY,
-  GRUNNLEGGENDE_BY_SLUG_QUERY,
   BLOGG_BY_SLUG_QUERY,
-  MONSTER_MALER_BY_SLUG_QUERY,
-  METADATA_BY_SLUG_QUERY,
-  BLOGG_LANDINGSSIDE_PAGE_QUERY,
   BLOGG_LANDINGSSIDE_BLOGS_QUERY,
+  BLOGG_LANDINGSSIDE_PAGE_QUERY,
+  DESIGNSYSTEM_GRUNNLEGGENDE_LANDINGPAGE_QUERY,
+  DESIGNSYSTEM_KOMPONENTER_LANDINGPAGE_QUERY,
+  DESIGNSYSTEM_OVERVIEW_BY_CATEGORY_QUERY,
+  DESIGNSYSTEM_OVERVIEW_PAGES_QUERY,
+  DESIGNSYSTEM_SIDEBAR_QUERY,
+  DESIGNSYSTEM_TEMPLATES_LANDINGPAGE_QUERY,
+  DOCUMENT_BY_ID_FOR_SLACK_QUERY,
+  GLOBAL_SEARCH_QUERY_ALL,
   GOD_PRAKSIS_ALL_TEMA_QUERY,
+  GOD_PRAKSIS_ARTICLE_BY_SLUG_QUERY,
+  GOD_PRAKSIS_ARTICLES_BY_TEMA_QUERY,
   GOD_PRAKSIS_LANDING_PAGE_SEO_QUERY,
   GOD_PRAKSIS_TEMA_BY_SLUG_QUERY,
-  GOD_PRAKSIS_ARTICLES_BY_TEMA_QUERY,
-  GOD_PRAKSIS_ARTICLE_BY_SLUG_QUERY,
-  DOCUMENT_BY_ID_FOR_SLACK_QUERY,
-  SIDE_ARTICLE_BY_SLUG_QUERY,
+  GRUNNLEGGENDE_BY_SLUG_QUERY,
+  KOMPONENT_BY_SLUG_QUERY,
+  METADATA_BY_SLUG_QUERY,
+  MONSTER_MALER_BY_SLUG_QUERY,
   PRINSIPPER_BY_SLUG_QUERY,
+  SIDE_ARTICLE_BY_SLUG_QUERY,
+  SITEMAP_ARTICLES_BY_TYPE_QUERY,
+  SITEMAP_LANDINGPAGES_QUERY,
+  SLUG_BY_TYPE_QUERY,
+  TOC_BY_SLUG_QUERY,
 };