+++ aksel.nav.no/website/app/dev/(designsystemet)/_ui/DesignsystemetPage.tsx
@@ -1,10 +1,12 @@
 import { PortableTextBlock } from "next-sanity";
+import { Image } from "sanity";
 import { BodyShort, Box, HStack, Heading, Tag } from "@navikt/ds-react";
 import {
   GRUNNLEGGENDE_BY_SLUG_QUERYResult,
   KOMPONENT_BY_SLUG_QUERYResult,
   MONSTER_MALER_BY_SLUG_QUERYResult,
 } from "@/app/_sanity/query-types";
+import { urlForImage } from "@/app/_sanity/utils";
 import { CustomPortableText } from "@/app/_ui/portable-text/CustomPortableText";
 import { getStatusTag } from "@/app/_ui/theming/theme-config";
 import { DesignsystemetEyebrow } from "@/app/dev/(designsystemet)/_ui/Designsystemet.eyebrow";
@@ -51,12 +53,19 @@ async function DesignsystemetPageHeader({ data }: DesignsystemetPageT) {
 
   const isComponentPage = data?._type === "komponent_artikkel";
 
+  const imageUrl = urlForImage(data?.status?.bilde as Image)?.url();
+
   return (
-    <Box marginBlock="space-0 space-28">
+    <Box marginBlock="space-0 space-28" data-color-role={statusTag?.colorRole}>
       <DesignsystemetEyebrow type={data?._type} />
 
       <Box marginBlock="space-0 space-8" asChild>
-        <Heading level="1" size="xlarge" data-aksel-heading-color>
+        <Heading
+          level="1"
+          size="xlarge"
+          data-aksel-heading-color
+          data-text-prose
+        >
           {data?.heading}
         </Heading>
       </Box>
@@ -87,7 +96,7 @@ async function DesignsystemetPageHeader({ data }: DesignsystemetPageT) {
       </HStack>
       {isComponentPage && <KomponentLinks data={data} />}
 
-      <DesignsystemetThumbnail />
+      <DesignsystemetThumbnail thumbnailUrl={imageUrl} />
     </Box>
   );
 }