+++ aksel.nav.no/website/pages/templates/skjemavalidering/react-hook-form.tsx
@@ -77,8 +77,6 @@ const Example = () => {
             })}
           />
           <RadioGroup
-            id="transportmiddel"
-            tabIndex={-1}
             legend="Transportmiddel"
             error={errors.transportmiddel?.message}
             onChange={() => trigger("transportmiddel")}
@@ -87,6 +85,7 @@ const Example = () => {
             {["Bil", "Gange", "Kollektivtransport"].map((value) => (
               <Radio
                 key={value}
+                id={value === "Bil" ? "transportmiddel" : undefined}
                 value={value}
                 {...register("transportmiddel", {
                   required: "Du må velge et transportmiddel.",